:recycle: refactor migrate `code-utils.js` to `src/shared`
