+++ api/tests/unit/infrastructure/utils/code-utils_test.js
@@ -1,4 +1,4 @@
-import { generateNumericalString } from '../../../../lib/infrastructure/utils/code-utils.js';
+import { generateNumericalString } from '../../../../src/shared/infrastructure/utils/code-utils.js';
 import { expect } from '../../../test-helper.js';
 
 describe('Unit | Utils | code-utils', function () {
